+++ src/util.js
@@ -1,0 +1,5 @@
+export const isObservable = (response$) =>
+  typeof response$.subscribe === 'function'
+
+export const isPromise = (response$) =>
+  typeof response$.then === 'function'